atualizações em desenvolvimento fluxo cadastro de pet

+++ app/Http/Controllers/Api/BlogPostagemController.php
@@ -71,14 +71,9 @@ class BlogPostagemController extends Controller
     public function storePostagem(Request $request)
     {
         $valida_titulo = BlogPostagem::where('titulo', $request->titulo)->first();
-        $valida_subtitulo = BlogPostagem::where('subtitulo', $request->subtitulo)->first();
 
         if ($valida_titulo != null) {
             return Response(['message' => 'Título já cadastrado'], Response::HTTP_CONFLICT);
-        }
-
-        if ($valida_subtitulo != null) {
-            return Response(['message' => 'Subtítulo já cadastrado'], Response::HTTP_CONFLICT);
         }
 
         $userAuth = Auth::user();
@@ -101,7 +96,7 @@ class BlogPostagemController extends Controller
             'imagem' => $caminhoImagem,
         ]);
 
-        if($request->tags !== null && !empty($request->tags)) {
+        if ($request->tags !== null && !empty($request->tags)) {
             $tags = $request->tags;
             $tags = BlogTag::whereIn('tag', $tags)->get();
             $postagem->tags()->attach($tags);
@@ -161,7 +156,7 @@ class BlogPostagemController extends Controller
     {
         $postagem = BlogPostagem::where('slug', $slug)->first();
 
-        if($postagem == null) {
+        if ($postagem == null) {
             return Response(['message' => 'Postagem não encontrada'], Response::HTTP_NOT_FOUND);
         }
 
@@ -179,7 +174,7 @@ class BlogPostagemController extends Controller
     {
         $postagem = BlogPostagem::where('slug', $slug)->first();
 
-        if($postagem == null) {
+        if ($postagem == null) {
             return Response(['message' => 'Postagem não encontrada'], Response::HTTP_NOT_FOUND);
         }
 
@@ -209,15 +204,10 @@ class BlogPostagemController extends Controller
      */
     public function updatePostagem(Request $request, string $id)
     {
-        $valida_titulo = BlogPostagem::where('titulo', $request->titulo)->first();
-        $valida_subtitulo = BlogPostagem::where('subtitulo', $request->subtitulo)->first();
+        $valida_request = BlogPostagem::where('titulo', $request->titulo)->first();
 
-        if ($valida_titulo != null) {
+        if ($valida_request != null && $valida_request->id != $id) {
             return Response(['message' => 'Título já cadastrado'], Response::HTTP_CONFLICT);
-        }
-
-        if ($valida_subtitulo != null) {
-            return Response(['message' => 'Subtítulo já cadastrado'], Response::HTTP_CONFLICT);
         }
 
         $postagem = BlogPostagem::find($id);
@@ -243,15 +233,16 @@ class BlogPostagemController extends Controller
             'titulo' => $request->titulo,
             'subtitulo' => $request->subtitulo,
             'conteudo' => $request->conteudo,
-            'flg_ativo' => $request->flg_ativo,
             'slug' => $slug,
             'imagem' => $caminhoImagem,
         ]);
 
-        $postagem->tags()->detach(BlogPostagemTag::where('blog_postagens_id', $postagem->id)->pluck('blog_tags_id'));
-        $tags = $request->tags;
-        $tags = BlogTag::whereIn('tag', $tags)->get();
-        $postagem->tags()->attach($tags);
+        if ($request->tags != null) {
+            $postagem->tags()->detach(BlogPostagemTag::where('blog_postagens_id', $postagem->id)->pluck('blog_tags_id'));
+            $tags = $request->tags;
+            $tags = BlogTag::whereIn('tag', $tags)->get();
+            $postagem->tags()->attach($tags);
+        }
 
         return Response(['message' => 'Postagem atualizada com sucesso'], Response::HTTP_OK);
     }
@@ -264,7 +255,7 @@ class BlogPostagemController extends Controller
         $postagem = BlogPostagem::find($id);
         $userAuth = Auth::user();
 
-        if($postagem == null) {
+        if ($postagem == null) {
             return Response(['message' => 'Postagem não encontrada'], Response::HTTP_NOT_FOUND);
         }
 
@@ -274,13 +265,13 @@ class BlogPostagemController extends Controller
 
         $tags = BlogPostagemTag::where('blog_postagens_id', $id);
 
-        if(!empty($tags)) {
+        if (!empty($tags)) {
             BlogPostagemTag::where('blog_postagens_id', $id)->delete();
         }
 
         $favoritada = BlogPostagemFavorita::where('blog_postagem_id', $id);
 
-        if($favoritada !== null) {
+        if ($favoritada !== null) {
             BlogPostagemFavorita::where('blog_postagem_id', $id)->delete();
         }
 
@@ -288,25 +279,25 @@ class BlogPostagemController extends Controller
         if ($caminhoImagemPostagem !== "api/imagens/blog/placeholder-blog.jpg" && File::exists($caminhoImagemPostagem)) {
             File::delete($caminhoImagemPostagem);
         }
+
         $postagem->delete();
 
-        return Response(['message' => 'Postagem foi removida com sucesso'], Response::HTTP_OK);
-
+        return Response(['message' => 'Postagem removida com sucesso'], Response::HTTP_OK);
     }
 
     public function postagensCadastradasUser()
     {
         $user = Auth::user();
 
-        $postagens = DB::table('blog_postagens')
-            ->where('user_id', $user->id)
+        $postagens = BlogPostagem::where('user_id', $user->id)
+            ->with('tags')
             ->paginate(Constants::REGISTROS_PAGINACAO);
 
         $tags = BlogTag::all();
 
         return $postagens->isEmpty()
-        ? Response(['message' => 'Nenhuma postagem cadastrada'], Response::HTTP_NOT_FOUND)
-        : Response(['postagens' => $postagens, 'tags' => $tags], Response::HTTP_OK);
+            ? Response(['message' => 'Nenhuma postagem cadastrada'], Response::HTTP_NOT_FOUND)
+            : Response(['postagens' => $postagens, 'tags' => $tags], Response::HTTP_OK);
     }
 
     public function postagensFavoritasUser()
@@ -325,7 +316,8 @@ class BlogPostagemController extends Controller
         return $postagens->isEmpty() ? Response(['message' => 'Nenhuma postagem favoritada'], Response::HTTP_NOT_FOUND) : Response($postagens, Response::HTTP_OK);
     }
 
-    public function favoritarPostagem(string $id) {
+    public function favoritarPostagem(string $id)
+    {
         $postagem = BlogPostagem::find($id);
         $user = Auth::user();
 
@@ -334,15 +326,15 @@ class BlogPostagemController extends Controller
         }
 
         $listIdPostagensFavoritadas = DB::table('blog_postagens_favoritas')
-        ->where('user_id', $user->id)
-        ->where('blog_postagem_id', $id)
-        ->pluck('blog_postagem_id');
+            ->where('user_id', $user->id)
+            ->where('blog_postagem_id', $id)
+            ->pluck('blog_postagem_id');
 
-    $postagemFavoritadoIsAtivo = DB::table('blog_postagens_favoritas')
-        ->where('user_id', $user->id)
-        ->where('blog_postagem_id', $id)
-        ->get()
-        ->first();
+        $postagemFavoritadoIsAtivo = DB::table('blog_postagens_favoritas')
+            ->where('user_id', $user->id)
+            ->where('blog_postagem_id', $id)
+            ->get()
+            ->first();
 
         if ($listIdPostagensFavoritadas->contains($id)) {
             if ($postagemFavoritadoIsAtivo->flg_ativo == 1) {
@@ -359,7 +351,7 @@ class BlogPostagemController extends Controller
                     ->get()
                     ->first();
 
-                return Response(['message' => 'Postagem foi favoritada com sucesso', 'postagem' => $postagemFavoritada], Response::HTTP_OK);
+                return Response(['message' => 'Postagem favoritada com sucesso', 'postagem' => $postagemFavoritada], Response::HTTP_OK);
             }
         }
 
@@ -368,11 +360,11 @@ class BlogPostagemController extends Controller
             'blog_postagem_id' => $postagem->id
         ]);
 
-        return Response(['message' => 'Postagem foi favoritada com sucesso', 'postagem' => $postagemFavoritada], Response::HTTP_OK);
-
+        return Response(['message' => 'Postagem favoritada com sucesso', 'postagem' => $postagemFavoritada], Response::HTTP_OK);
     }
 
-    public function desfavoritarPostagem(string $id) {
+    public function desfavoritarPostagem(string $id)
+    {
         $postagem = BlogPostagem::find($id);
         $user = Auth::user();
 
@@ -381,9 +373,9 @@ class BlogPostagemController extends Controller
         }
 
         $listIdPostagensFavoritadas = DB::table('blog_postagens_favoritas')
-        ->where('user_id', $user->id)
-        ->where('blog_postagem_id', $id)
-        ->pluck('blog_postagem_id');
+            ->where('user_id', $user->id)
+            ->where('blog_postagem_id', $id)
+            ->pluck('blog_postagem_id');
 
         $postagemDesfavoritadoIsAtivo = DB::table('blog_postagens_favoritas')
             ->where('user_id', $user->id)
@@ -391,24 +383,24 @@ class BlogPostagemController extends Controller
             ->get()
             ->first();
 
-            if ($listIdPostagensFavoritadas->contains($id)) {
-                if ($postagemDesfavoritadoIsAtivo->flg_ativo == 0) {
-                    return Response(['message' => 'Postagem já desfavoritada'], Response::HTTP_CONFLICT);
-                } else {
-                    DB::table('blog_postagens_favoritas')
-                        ->where('user_id', $user->id)
-                        ->where('blog_postagem_id', $id)
-                        ->update(['flg_ativo' => 0]);
+        if ($listIdPostagensFavoritadas->contains($id)) {
+            if ($postagemDesfavoritadoIsAtivo->flg_ativo == 0) {
+                return Response(['message' => 'Postagem já desfavoritada'], Response::HTTP_CONFLICT);
+            } else {
+                DB::table('blog_postagens_favoritas')
+                    ->where('user_id', $user->id)
+                    ->where('blog_postagem_id', $id)
+                    ->update(['flg_ativo' => 0]);
 
-                    $postagemDesfavoritada = DB::table('blog_postagens_favoritas')
-                        ->where('user_id', $user->id)
-                        ->where('blog_postagem_id', $id)
-                        ->get()
-                        ->first();
+                $postagemDesfavoritada = DB::table('blog_postagens_favoritas')
+                    ->where('user_id', $user->id)
+                    ->where('blog_postagem_id', $id)
+                    ->get()
+                    ->first();
 
-                    return Response(['message' => 'Postagem foi desfavoritada com sucesso', 'postagem' => $postagemDesfavoritada], Response::HTTP_OK);
-                }
+                return Response(['message' => 'Postagem desfavoritada com sucesso', 'postagem' => $postagemDesfavoritada], Response::HTTP_OK);
             }
+        }
 
         DB::table('blog_postagens_favoritas')
             ->where('user_id', $user->id)
@@ -421,6 +413,6 @@ class BlogPostagemController extends Controller
             ->get()
             ->first();
 
-            return Response(['message' => 'Postagem foi desfavoritada com sucesso', 'postagem' => $postagemDesfavoritada], Response::HTTP_OK);
+        return Response(['message' => 'Postagem desfavoritada com sucesso', 'postagem' => $postagemDesfavoritada], Response::HTTP_OK);
     }
 }
